Fix renaming typos
