+++ x-pack/plugins/fleet/public/applications/fleet/sections/agent_policy/components/actions_menu.tsx
@@ -9,7 +9,7 @@ import { EuiContextMenuItem, EuiPortal } from '@elastic/eui';
 import { AgentPolicy } from '../../../types';
 import { useCapabilities } from '../../../hooks';
 import { ContextMenuActions } from '../../../components';
-import { AgentEnrollmentFlyout } from '../../fleet/components';
+import { AgentEnrollmentFlyout } from '../../agents/components';
 import { AgentPolicyYamlFlyout } from './agent_policy_yaml_flyout';
 import { AgentPolicyCopyProvider } from './agent_policy_copy_provider';
